backend/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
passlib[bcrypt]
pymongo
python-dotenv
pydantic
email-validator
